Use proper crate organization.
Fixes #14

Review: @bemeurer
